add example and demo for sdk and update Catalog and SDK version
