+++ docs/catalog/example.md
@@ -240,7 +240,7 @@ const BuyAsset = ({ddo}: {ddo: DDO}) => {
     }
 
     const currentAccount = await getCurrentAccount(sdk);
-    const response = await subscription.buySubscription(ddo.id, currentAccount, owner, 1, 1155);
+    const response = await subscription.buySubscription(ddo.id, currentAccount, owner, BigNumber.from(1), 1155);
     setIsBought(Boolean(response));
   };
 
@@ -309,7 +309,6 @@ const App = () => {
       author: '',
       license: '',
       dateCreated: new Date().toISOString(),
-      price: ''
     }
   };
 
@@ -317,7 +316,7 @@ const App = () => {
     try {
       const publisher = await getCurrentAccount(sdk);
       const rewardsRecipients: any[] = [];
-      const assetRewardsMap = constructRewardMap(rewardsRecipients, 100, publisher.getId());
+      const assetRewardsMap = constructRewardMap(rewardsRecipients, BigNumber.from(100), publisher.getId());
       const assetRewards = new AssetRewards(assetRewardsMap);
       const royaltyAttributes = {
         royaltyKind: RoyaltyKind.Standard,
@@ -332,9 +331,9 @@ const App = () => {
         gatewayAddress: String(appConfig.gatewayAddress),
         assetRewards,
         metadata,
-        nftAmount: 1,
+        nftAmount: BigNumber.from(1),
         preMint: true,
-        cap: 100,
+        cap: BigNumber.from(100),
         royaltyAttributes,
         erc20TokenAddress,
       });
@@ -479,7 +478,7 @@ const BuyAsset = ({ddo}: {ddo: DDO}) => {
     }
 
     const currentAccount = await getCurrentAccount(sdk);
-    const response = await subscription.buySubscription(ddo.id, currentAccount, owner, 1, 1155);
+    const response = await subscription.buySubscription(ddo.id, currentAccount, owner, BigNumber.from(1), 1155);
     setIsBought(Boolean(response));
   };
 
@@ -538,7 +537,6 @@ const App = () => {
       author: '',
       license: '',
       dateCreated: new Date().toISOString(),
-      price: ''
     }
   };
 
@@ -546,7 +544,7 @@ const App = () => {
     try {
       const publisher = await getCurrentAccount(sdk);
       const rewardsRecipients: any[] = [];
-      const assetRewardsMap = constructRewardMap(rewardsRecipients, 100, publisher.getId());
+      const assetRewardsMap = constructRewardMap(rewardsRecipients, BigNumber.from(100), publisher.getId());
       const assetRewards = new AssetRewards(assetRewardsMap);
       const royaltyAttributes = {
         royaltyKind: RoyaltyKind.Standard,
@@ -561,9 +559,9 @@ const App = () => {
         gatewayAddress: String(appConfig.gatewayAddress),
         assetRewards,
         metadata,
-        nftAmount: 1,
+        nftAmount: BigNumber.from(1),
         preMint: true,
-        cap: 100,
+        cap: BigNumber.from(100),
         royaltyAttributes,
         erc20TokenAddress,
       });
