+++ package.json
@@ -19,18 +19,20 @@
     "@docusaurus/core": "2.1.0",
     "@docusaurus/plugin-google-analytics": "2.1.0",
     "@docusaurus/preset-classic": "2.1.0",
-    "@docusaurus/theme-search-algolia": "2.1.0",
     "@docusaurus/theme-classic": "2.1.0",
     "@docusaurus/theme-common": "2.1.0",
+    "@docusaurus/theme-search-algolia": "2.1.0",
     "@mdx-js/react": "^1.6.22",
-    "@nevermined-io/catalog-core": "0.2.1",
-    "@nevermined-io/catalog-providers": "0.2.1",
+    "@nevermined-io/catalog-core": "0.3.0",
+    "@nevermined-io/catalog-providers": "0.3.0",
+    "@nevermined-io/nevermined-sdk-js": "^0.25.0",
     "@nevermined-io/styles": "^0.1.1",
     "axios": "^0.27.2",
     "bem-helpers": "^1.1.0",
     "clsx": "^1.1.1",
     "docusaurus-plugin-sass": "^0.2.2",
     "ethers": "^5.7.1",
+    "jose": "^4.10.0",
     "prism-react-renderer": "^1.3.3",
     "promise-polyfill": "8.2.3",
     "react": "^17.0.2",
