+++ src/components/demo/styles.module.scss
@@ -8,6 +8,10 @@
     border: 1px solid #B1D3EA;
   }
 
+  &__mint {
+    margin-left: 10px;
+  }
+
   &__detail {
     margin-right: 5px;
     color: #000;
